requirements.txt
Django>=2.0,<2.1
wagtail
psycopg2
psycopg2_binary
django-allauth
